moved rotate to utils.cpp
