+++ presets.hpp
@@ -2,5 +2,4 @@
 #pragma once
 #include <armadillo>
 
-arma::umat rotate(arma::umat& mat);
 extern arma::umat presets[10];
