+++ utils.cpp
@@ -7,3 +7,15 @@ bool isValidMousePosition(sf::Event::MouseButtonEvent mouseEvent, sf::Vector2u s
 
     return validX && validY;
 }
+
+arma::umat rotate(arma::umat& mat)
+{
+    arma::umat rotated(mat.n_cols, mat.n_rows);
+
+    for (int i = 0; i < mat.n_rows; i++)
+    {
+        rotated.col(i) = mat.row(mat.n_rows - i - 1).as_col();
+    }
+    
+    return rotated;
+}
